# .lev/pm/specs/deadzone-api-contract.yaml
openapi: 3.1.0
info:
  title: DeadZone MVP API
  version: 0.1.0
  description: Local-first FastAPI contract for deterministic mock/replay venue intelligence.
servers:
  - url: http://localhost:8000
    description: Local backend
paths:
...
  /api/v1/replay/control:
    post:
      summary: Load, play, pause, restart, seek, or speed-control replay mode.
      operationId: controlReplay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplayControlRequest"
      responses:
        "200":
          description: Updated stream state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamState"
        "409":
          description: Requested replay action conflicts with current mode or scenario state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/capture/start:
    post:
      summary: Start local BLE advertisement capture into a privacy-safe trace.
      operationId: startBleCapture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureStartRequest"
      responses:
        "200":
          description: Capture started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureStatus"
        "409":
          description: BLE dependency or permission is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureStatus"
...
components:
  schemas:
    Mode:
      type: string
      enum: [mock, replay, ble, wifi, mesh, hybrid]
...
    ErrorResponse:
      type: object
      required: [code, message, recoverable]
      properties:
        code:
          type: string
        message:
          type: string
        recoverable:
          type: boolean
        detail:
          type: object
          additionalProperties: true
...
    StreamState:
      type: object
      required: [mode, connected, sequence, source_label, freshness, clock]
      properties:
        mode:
          $ref: "#/components/schemas/Mode"
        connected:
          type: boolean
        sequence:
          type: integer
        source_label:
          type: string
          example: "Replay: local BLE trace ble-20260514-120000"
        freshness:
          type: string
          enum: [live, fresh, stale, offline]
        clock:
          type: object
          required: [timestamp, replay_position_s, speed]
          properties:
            timestamp:
              type: string
              format: date-time
            replay_position_s:
              type: number
            speed:
              type: number
        playback_state:
          type: string
          enum: [idle, playing, paused, ended]
          nullable: true
        scenario_id:
          type: string
          nullable: true
          description: Opaque replay scenario id for a locally captured BLE trace.
        duration_s:
          type: number
          nullable: true
...
    ReplayControlRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [load, play, pause, restart, seek, set_speed]
        scenario_id:
          type: string
        position_s:
          type: number
        speed:
          type: number
    CaptureStartRequest:
      type: object
      properties:
        trace_id:
          type: string
          description: Optional local operator label; API responses use opaque ids.
        duration_s:
          type: number
    CaptureStatus:
      type: object
      required: [active, observations, backend]
      properties:
        active:
          type: boolean
        trace_id:
          type: string
          nullable: true
        observations:
          type: integer
        backend:
          type: string
          const: ble
        disabled_reason:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        latest_observation_at:
          type: string
          format: date-time
          nullable: true